% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lm_annotation}
\alias{add_lm_annotation}
\title{Add linear model annotation to an existing plot}
\usage{
add_lm_annotation(
  fit,
  col = "darkred",
  cex = 1,
  x_offset = 0.05,
  y_offset = 0.05
)
}
\arguments{
\item{fit}{A fitted lm object}

\item{col}{Text color (default = "darkred")}

\item{cex}{Text size (default = 1)}

\item{x_offset}{Relative x offset from the left (default = 0.05)}

\item{y_offset}{Relative y offset from the top (default = 0.05)}
}
\description{
Add linear model annotation to an existing plot
}
