% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check that required columns are present in a data frame}
\usage{
check_columns(df, required_cols)
}
\arguments{
\item{df}{A data frame to check.}

\item{required_cols}{A character vector of required column names.}
}
\value{
Nothing.
}
\description{
Verifies that all expected column names are present in the input data frame.
}
