% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check file path and stop if it does not exist}
\usage{
check_file_exists(name, path)
}
\arguments{
\item{name}{Name of the file (for display purposes)}

\item{path}{Path to the file}
}
\description{
Check file path and stop if it does not exist
}
