% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_identical_snps}
\alias{check_identical_snps}
\title{Check that multiple SNP tables share identical SNP sets}
\usage{
check_identical_snps(dfs, context = NULL)
}
\arguments{
\item{dfs}{[list] A list of data frames or matrices to check. Each element
must have rownames corresponding to SNP identifiers (e.g.: all_freqs or read_depths dataframes)}

\item{context}{[character] Optional string describing the context of the check}
}
\value{
Invisibly returns `TRUE` if all tables contain identical SNP sets.
  Stops execution with an error message if any mismatch is detected.
}
\description{
Check that multiple SNP tables share identical SNP sets
}
