% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_files}
\alias{check_input_files}
\title{Check required and optional input file paths}
\usage{
check_input_files(required_files, optional_files = NULL)
}
\arguments{
\item{required_files}{Named list of required file/folder paths}

\item{optional_files}{Named list of optional file/folder paths (default: NULL)}
}
\description{
Check required and optional input file paths
}
