% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing_args}
\alias{check_missing_args}
\title{Check that all required arguments are provided}
\usage{
check_missing_args(args)
}
\arguments{
\item{args}{Named list of required arguments to check.
Each element must be non-NULL.}
}
\value{
Nothing. The function stops with an error if at least one element is NULL.
}
\description{
This function checks that all elements of a named list are non-NULL.
}
