% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_depth_file}
\alias{clean_depth_file}
\title{Clean depth file 2 first columns}
\usage{
clean_depth_file(depths)
}
\arguments{
\item{depths}{Read depth data frame (SNPs x samples) with CHROM and POS as first two columns.}
}
\value{
A cleaned read depth data frame with chrom_pos as row names and no CHROM/POS columns.
}
\description{
Clean depth file 2 first columns
}
