% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_boxplots_est_freq}
\alias{compare_boxplots_est_freq}
\title{Plot boxplots of (1 or more sets of) estimated frequencies by expected frequencies}
\usage{
compare_boxplots_est_freq(freqs_df, variable_name, out_file = NULL)
}
\arguments{
\item{freqs_df}{Dataframe with at least one estimated frequency column and a "ExpFreq" column (providing the expected frequencies)}

\item{variable_name}{Label for x-axis (below all conditions)}

\item{out_file}{Path to the output file to save the plot (default = NULL)}
}
\value{
A ggplot object
}
\description{
Plot boxplots of (1 or more sets of) estimated frequencies by expected frequencies
}
