% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_bias}
\alias{compute_bias}
\title{Compute estimation bias and plot grouped error boxplots}
\usage{
compute_bias(freqs_df, variable_name = "", out_dir = NULL, suffix = "")
}
\arguments{
\item{freqs_df}{Dataframe with an "ExpFreq" colomn (providing the expected frequencies), a "Component" column, and 1 or more estimated frequencies column(s)}

\item{variable_name}{Label for the x-axis (e.g. name of the strategy variable)}

\item{out_dir}{Path the the output directory to write output files (default = NULL)}

\item{suffix}{Suffix to add in the output file names (default = "")}
}
\value{
A list with bias dataframes per component and per expected frequency
}
\description{
Compute estimation bias and plot grouped error boxplots
}
