% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_errors}
\alias{compute_errors}
\title{Compute estimation errors}
\usage{
compute_errors(freqs_df, out_dir = NULL, suffix = "")
}
\arguments{
\item{freqs_df}{Dataframe with at least an "ExpFreq" colomn (providing the expected frequencies) and an estimated frequencies column}

\item{out_dir}{Path the the output directory to write output files (default = NULL)}

\item{suffix}{Suffix to add in the output file names (default = "")}
}
\value{
An error dataframe
}
\description{
Compute estimation errors
}
