% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mean_depth_per_marker}
\alias{compute_mean_depth_per_marker}
\title{Compute and mean depth per marker}
\usage{
compute_mean_depth_per_marker(depths, out_file = NULL)
}
\arguments{
\item{depths}{Read depth data frame (SNPs x samples) with CHROM and POS as first two columns.}

\item{out_file}{Optional Path to save the output as a TSV.}
}
\value{
A data frame of mean depths per marker (printed or saved).
}
\description{
Compute and mean depth per marker
}
