% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_stats_per_exp_freq}
\alias{compute_stats_per_exp_freq}
\title{Compute mean and SD per expected frequency}
\usage{
compute_stats_per_exp_freq(freqs_df, out_dir = NULL, suffix = "")
}
\arguments{
\item{freqs_df}{Dataframe with an "ExpFreq" colomn (providing the expected frequencies) and 1 or more estimated frequencies column(s)}

\item{out_dir}{Path the the output directory to write output files (default = NULL)}

\item{suffix}{Suffix to add in the output file names (default = "")}
}
\value{
A list with two dataframes: expected frequencies mean and standard deviation
}
\description{
Compute mean and SD per expected frequency
}
