% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_freq_errors_subsampling}
\alias{estimate_freq_errors_subsampling}
\title{Estimate genotype frequency errors under SNP subsampling with replicates}
\usage{
estimate_freq_errors_subsampling(
  genotyping_matrix,
  allele_freqs,
  step_size,
  nb_reps = 5,
  snp_depths = NULL,
  expected_freqs_melt
)
}
\arguments{
\item{genotyping_matrix}{Genotyping matrix with values 1 (hom. REF), 0.5 (het), and 0 (hom. ALT) (SNPs x components)}

\item{allele_freqs}{Allele frequency matrix (SNPs x mixtures)}

\item{step_size}{Number of SNPs to increment by when subsampling}

\item{nb_reps}{Integer. Number of random subsampling replicates per SNP subset size (default = 5).}

\item{snp_depths}{Read depths matrix (SNPs x mixtures) (default = NULL)}

\item{expected_freqs_melt}{Melted dataframe of expected genotype frequencies}
}
\value{
A dataframe of the estimation errors
}
\description{
Estimate genotype frequency errors under SNP subsampling with replicates
}
