% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_geno_freqs_snps_subsampling}
\alias{estimate_geno_freqs_snps_subsampling}
\title{Compare the effect of random SNP subsampling on genotype frequency estimation}
\usage{
estimate_geno_freqs_snps_subsampling(
  genotyping_matrix,
  allele_freqs,
  step_size,
  snp_depths = NULL,
  expected_freqs_melt = NULL
)
}
\arguments{
\item{genotyping_matrix}{Genotyping matrix with values 1 (hom. REF), 0.5 (het), and 0 (hom. ALT) (SNPs x components)}

\item{allele_freqs}{Allele frequency matrix (SNPs x mixtures)}

\item{step_size}{Number of SNPs to increment by when subsampling}

\item{snp_depths}{Read depths matrix (SNPs x mixtures) (default = NULL)}

\item{expected_freqs_melt}{Melted dataframe of expected genotype frequencies (default = NULL)}
}
\value{
Dataframe of estimated frequencies per condition
}
\description{
Compare the effect of random SNP subsampling on genotype frequency estimation
}
