% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_genotype_freqs}
\alias{estimate_genotype_freqs}
\title{Estimate genotype frequencies from genotyping matrix and allele frequencies}
\usage{
estimate_genotype_freqs(
  genotyping_matrix,
  allele_freqs,
  snp_depths = NULL,
  expected_freqs_melt = NULL,
  est_freq_col_name = "EstFreq",
  out_file = NULL
)
}
\arguments{
\item{genotyping_matrix}{Genotyping matrix with values 1 (hom. REF), 0.5 (het), and 0 (hom. ALT) (SNPs x components)}

\item{allele_freqs}{Allele frequency matrix (SNPs x mixtures)}

\item{snp_depths}{Read depths matrix (SNPs x mixtures) (default = NULL)}

\item{expected_freqs_melt}{Melted dataframe of expected genotype frequencies (default = NULL)}

\item{est_freq_col_name}{Column name for estimated frequencies (default = "EstFreq")}

\item{out_file}{Path to the output file to write the estimated genotype frequencies (default = NULL)}
}
\value{
A melted dataframe of the estimated genotype frequencies
}
\description{
Estimate genotype frequencies from genotyping matrix and allele frequencies
}
