% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_pipeline}
\alias{estimate_pipeline}
\title{Estimate genotype frequencies pipeline}
\usage{
estimate_pipeline(
  genotyping_vcf_path,
  allele_freqs_path,
  snp_depths_path,
  lib_names_corresp_path = NULL,
  min_depth = 0,
  output_file_name
)
}
\arguments{
\item{genotyping_vcf_path}{Path to genotyping VCF}

\item{allele_freqs_path}{Path to allele frequencies file}

\item{snp_depths_path}{Path to SNP depths file}

\item{lib_names_corresp_path}{Optional path to library name correspondence file}

\item{min_depth}{Minimum read depth threshold}

\item{output_file_name}{Output file path}
}
\value{
Invisibly returns the output file path
}
\description{
This function estimates genotype frequencies from a genotyping
VCF, allele frequencies and SNP depths.
}
