% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_weight_vector_effect}
\alias{evaluate_weight_vector_effect}
\title{Evaluate the effect of weight vector on genotype frequency estimation}
\usage{
evaluate_weight_vector_effect(
  genotyping_matrix,
  allele_freqs,
  snp_depths,
  expected_freqs_melt,
  out_dir
)
}
\arguments{
\item{genotyping_matrix}{Genotyping matrix with values 1 (hom. REF), 0.5 (het), and 0 (hom. ALT) (SNPs x components)}

\item{allele_freqs}{Allele frequency matrix (SNPs x mixtures)}

\item{snp_depths}{Read depths matrix (SNPs x mixtures)}

\item{expected_freqs_melt}{Melted dataframe of expected genotype frequencies}

\item{out_dir}{Path the the output directory to write output files}
}
\value{
A named list with two elements:
\describe{
  \item{stats}{A list of dataframes from \code{compute_stats_per_exp_freq()}: mean and standard deviation per expected frequency}
  \item{bias}{A list of dataframes from \code{compute_bias()}: bias per component and per expected frequency}
}
}
\description{
This function compares estimated genotype frequencies using:
- read depth as weight
- no weighting
Then compares both results with plots, bias, and summary stats.
}
