% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_alphabetical}
\alias{factor_alphabetical}
\title{Force alphabetical ordering of a column based on labels}
\usage{
factor_alphabetical(x)
}
\arguments{
\item{x}{A vector (character or factor).}
}
\value{
A factor with levels sorted alphabetically.
}
\description{
Force alphabetical ordering of a column based on labels
}
