% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MAF_hist_pipeline.R
\name{MAF_hist_pipeline}
\alias{MAF_hist_pipeline}
\title{Minor allele frequency analysis pipeline}
\usage{
MAF_hist_pipeline(genotyping_vcf_path, lib_names_corresp_path = NULL, out_dir)
}
\arguments{
\item{genotyping_vcf_path}{Path to genotyping VCF}

\item{lib_names_corresp_path}{Optional path to library name correspondence file}

\item{out_dir}{Output directory}
}
\value{
Invisibly returns the output file path
}
\description{
Run the complete pipeline to compute and plot minor allele frequency (MAF)
distributions and marker set intersections from a genotyping VCF.
}
