% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf_pipeline}
\alias{maf_pipeline}
\title{Minor allele frequency analysis pipeline}
\usage{
maf_pipeline(genotyping_vcf_path, lib_names_corresp_path = NULL, out_dir)
}
\arguments{
\item{genotyping_vcf_path}{Path to genotyping VCF}

\item{lib_names_corresp_path}{Optional path to library name correspondence file}

\item{out_dir}{Output directory}
}
\value{
Invisibly returns the output file path
}
\description{
This function computes and plots minor allele frequency (MAF)
distributions and marker set intersections from a genotyping VCF.
}
