% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{melt_genotype_freqs}
\alias{melt_genotype_freqs}
\title{Melt a Genotype Frequency Table to Long Format}
\usage{
melt_genotype_freqs(genotype_freqs, freq_col_name)
}
\arguments{
\item{genotype_freqs}{Dataframe containing genotype frequencies, with a "Mixture" column and at least two other columns representing components.}

\item{freq_col_name}{A character string specifying the name to use for the frequency column in the output.}
}
\value{
A dataframe in long format, with columns Mixture, Component, and a frequency column named according to "freq_col_name".
}
\description{
Melt a Genotype Frequency Table to Long Format
}
