% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_depth_files}
\alias{merge_depth_files}
\title{Merge multiple depth files by CHROM and POS}
\usage{
merge_depth_files(depth_list)
}
\arguments{
\item{depth_list}{Path to a text file listing depth file paths (one per line)
Listed files must be read depths matrices (SNPs x samples) with CHROM and POS as first and second columns.}
}
\value{
A merged data.frame containing all depth files joined by CHROM and POS.
}
\description{
Merge multiple depth files by CHROM and POS
}
