% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_gt_matrix_to_majmin}
\alias{num_gt_matrix_to_majmin}
\title{Reorient REF/ALT-coded genotype matrix to major/minor per SNP}
\usage{
num_gt_matrix_to_majmin(numeric_matrix)
}
\arguments{
\item{numeric_matrix}{Numeric matrix/data.frame of genotypes (rows = SNPs, cols = samples):
1 = hom. REF, 0.5 = het, 0 = hom. ALT.}
}
\value{
Numeric matrix with rows reoriented to major/minor:
  1 = hom. major, 0.5 = het, 0 = hom. minor.
}
\description{
Reorient REF/ALT-coded genotype matrix to major/minor per SNP
}
