% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_condition_effect_boxplots}
\alias{plot_condition_effect_boxplots}
\title{Plot boxplots across varying conditions}
\usage{
plot_condition_effect_boxplots(
  df_melt,
  x_col,
  y_col,
  fill_col,
  ref_col = NULL,
  y_lab = "",
  x_lab = "",
  fill_lab = "Group",
  show_zero = F
)
}
\arguments{
\item{df_melt}{Data frame in long format}

\item{x_col}{Column name for x-axis (string)}

\item{y_col}{Column name for y-axis (string)}

\item{fill_col}{Column name for fill (string)}

\item{ref_col}{Optional column to use for horizontal dashed lines (default = NULL)}

\item{y_lab}{Label for y-axis}

\item{x_lab}{Label for x-axis}

\item{fill_lab}{Legend title for fill}

\item{show_zero}{Logical (default = FALSE). If TRUE, adds a horizontal dashed line at y = 0 to help visualize deviations around zero.}
}
\value{
A ggplot object
}
\description{
Plot boxplots across varying conditions
}
