% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_correlation_with_exp_freq}
\alias{plot_correlation_with_exp_freq}
\title{Plot estimated vs expected frequencies with regression}
\usage{
plot_correlation_with_exp_freq(
  freqs_df,
  extra_freqs_df = NULL,
  out_file = NULL
)
}
\arguments{
\item{freqs_df}{Dataframe with an "ExpFreq" colomn (providing the expected frequencies) and 1 or more estimated frequencies column(s)}

\item{extra_freqs_df}{Dataframe with additional points to plot (they won't be used to compute the regression) (default = NULL)}

\item{out_file}{Path to the output file to save the plot (default = NULL)}
}
\description{
Plot estimated vs expected frequencies with regression
}
