% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_error_boxplots}
\alias{plot_error_boxplots}
\title{Plot grouped error boxplots by component and expected frequency}
\usage{
plot_error_boxplots(errors_long_df, variable_name, out_dir = NULL, suffix = "")
}
\arguments{
\item{errors_long_df}{Long-format dataframe with columns: Component, ExpFreq, Error, Condition}

\item{variable_name}{Label to use for the x-axis (i.e., the condition name shown on plots).}

\item{out_dir}{Optional output directory to save plots (default = NULL)}

\item{suffix}{Optional suffix to append to output file names (default = "")}
}
\value{
NULL. Plots are printed if no output directory is given.
}
\description{
Plot grouped error boxplots by component and expected frequency
}
