% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_MAF_hist}
\alias{plot_MAF_hist}
\title{Plot histogram of Minor Allele Frequencies (MAF) from a VCF object}
\usage{
plot_MAF_hist(vcf, x_lim_values, out_dir = NULL)
}
\arguments{
\item{vcf}{A `vcfR` object, as read by [vcfR::read.vcfR()].}

\item{x_lim_values}{A numeric vector of length 2 indicating the minimum and maximum x-axis limits (e.g., `c(0, 0.5)`).}
}
\value{
A histogram plot of the MAF distribution. No return value; the function is used for its side effect (plot).
}
\description{
This function extracts the minor allele frequencies (MAF) from a `vcfR` object
and plots a histogram of their distribution.
}
