% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_marker_set_intersections}
\alias{plot_marker_set_intersections}
\title{Plot marker set intersections as an UpSet and save to PNG.}
\usage{
plot_marker_set_intersections(numeric_matrix, out_file = NULL)
}
\arguments{
\item{numeric_matrix}{A numeric genotyping matrix (rows = SNPs, columns = samples) where 1 = homozygous for the major allele, 0 = homozygous for the minor allele. Typically produced by [vcf_to_majmin_numeric_matrix()].}

\item{out_file}{Output PNG file path.}
}
\value{
An UpSet plot of the marker set intersections. No return value; the function is used for its side effect (plot).
}
\description{
Plot marker set intersections as an UpSet and save to PNG.
}
