% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_rownames_from_chrom_pos}
\alias{set_rownames_from_chrom_pos}
\title{Set row names from CHROM and POS columns and remove them}
\usage{
set_rownames_from_chrom_pos(df)
}
\arguments{
\item{df}{A data frame with columns named `CHROM` and `POS`.}
}
\value{
A data frame with updated row names and without the `CHROM` and `POS` columns.
}
\description{
Combines the `CHROM` and `POS` columns to form row names of the form `CHROM_POS`,
and removes those two columns from the data frame.
}
