% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcf_to_majmin_num_gt_matrix}
\alias{vcf_to_majmin_num_gt_matrix}
\title{Build a major/minor-oriented numeric genotype matrix from a VCF}
\usage{
vcf_to_majmin_num_gt_matrix(vcf_path, lib_names_corresp_path = NULL)
}
\arguments{
\item{vcf_path}{Path to VCF file.}

\item{lib_names_corresp_path}{Optional path to a two-column mapping table
(Library_name -> Genotype) for renaming samples.}
}
\value{
Numeric genotyping matrix (SNPs x samples) oriented to major/minor:
  1 = hom. major, 0.5 = het, 0 = hom. minor.
}
\description{
Build a major/minor-oriented numeric genotype matrix from a VCF
}
