% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcf_to_numeric_matrix}
\alias{vcf_to_numeric_matrix}
\title{Extract genotype matrix from VCF and convert to numeric}
\usage{
vcf_to_numeric_matrix(vcf_path, lib_names_corresp_path = NULL)
}
\arguments{
\item{vcf_path}{Path to VCF file}

\item{lib_names_corresp_path}{Path to a table matching sample library names to sample genotype names (default = NULL)}
}
\value{
Numeric genotyping matrix with values 1 (hom. REF), 0.5 (het), and 0 (hom. ALT) (SNPs x samples)
}
\description{
Extract genotype matrix from VCF and convert to numeric
}
